+++ changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - primal heuristics derived from `scip::ObjHeur` via `Model::includeHeur`,
   - node selectors derived from `scip::ObjNodesel` via `Model::includeNodesel`,
   - presolvers derived from `scip::ObjPresol` via `Model::includePresol`,
+  - pricers derived from `scip::ObjPricer` via `Model::includePricer`,
   - propagators derived from `scip::ObjProp` via `Model::includeProp`,
   - relaxators derived from `scip::ObjRelax` via `Model::includeRelax`,
   - separators derived from `scip::ObjSepa` via `Model::includeSepa`, and